requirements.txt
wikipedia-api
beautifulsoup4
requests
PyMuPDF
pdfplumber
ebooklib
sentence-transformers
faiss-cpu
numpy
openai==0.28
google-generativeai
